Added all callbacks for the File menu (except for ROM History).
Fixed a bug from the old GTK+ UI when selecting a savestate slot. In the
old UI, selecting a slot would cause Set_Current_State() to be run twice;
once with the previous slot, and once with the new slot. The callback
has been updated to only call Set_Current_State() if the current menu
item is checked.