Revert "Add a mnemonic with the entry number in the "ROM History" submenu."
This reverts commit 4da04303a351b6f03c3b92f386df7a4cac6318a8.

GTK+ apparently supports multiple mnemonics in a label. So, if a file like
scd_510.iso was in ROM History, it'd show up like this:

1. [SCD] scd510.iso

with both the first 1 and the 5 in scd_510.iso being mnemonics. This doesn't
work well here, but rather than waste time figuring out how to fix it, I'll
just revert it. I can fix it later on.