Workaround for a stupid bug in zlib.
zlib has a stupid bug where, if a file pointer was fread() somewhere
else in the code, it will refuse to recognize said file pointer as a
gzip file in gzdopen(), even if the file pointer was fseek()'d back
to the beginning of the file.

Workaround: Pass the filename to the COMPRESS_Get_First_File_Size()
and COMPRESS_Get_First_File() functions.