Partially reverted commit c6e201596fee028dd542d3a4482cc6c332f51e47.
Separating everything in gens/ into separate libraries doesn't work,
since the files are all interdependent on each other. Thus, if one
file is changed in one directory, it won't trigger a rebuild on
dependent files, causing problems.

This is only a partial revert because I left in some improvements
in the Makefile.am files, such as cleanups.