New and improved bitmap screenshot function.
Instead of getting the rendered image from SDL, the new screenshot
function gets the image from MD_Screen[]. Contrast and brightness
are preapplied to MD_Screen[], but filters are not.

The bitmap is saved as 320x224 (or 320x240 depending on VDP mode).

TODO: Save as 256x224 (or 256x240) in low-resolution mode.