Finally fixed 32-bit color on 32X!
It turns out that the problem was caused by the 32X asm code in vdp_rend.asm
not being disabled. The macros Upth gave me to do the 32X handling were
supposed to replace the asm code. Instead, both the asm and the C macros
were processing the video, which led to broken colors.

But it works now. :)