Compressor::getFile(): Fixed incorrect parameter ordering in fread().
Load_ROM() now checks if the filesize returned by getFile() is the
same as the filesize reported by the archive file. For uncompressed
ROMs, Compressor::getFile() was specifying the ROM size as the element
size and to read one of said element, so fread() returned 1 instead
of the number of bytes. Swapping these two parameters fixed the bug.