Added an easter egg. :D
This easter egg replaces the Gens logo in the About dialog with
an animated Super Sonic (S2) image.

To activate the easter egg, do the following:
1. Load Sonic 2, rev00.
2. Get all 7 Chaos Emeralds.
3. Pause Gens by pressing Esc.
4. Click Help, About.

WARNING: Do NOT do any of the following:
- Use cheat codes in-game.
- Open the Game Genie dialog.
- Have a pre-made .pat file for Sonic 2 with codes already set up.
- Save or load a savestate.

If any of the above are done, the easter egg is automatically
disabled. In order to re-enable the easter egg, you have to close
the ROM and reload it.