Added zip.c and zip.h from Minizip.
Nemesis is apparently working on a new savestate format that stores
various pieces of information as different files in a Zip archive,
so in order to prepare for that, I added a Zip compressor. It's
currently not compiled, but can easily be added to Makefile.am when
the savestate format is finished.