New UI_Sleep() function, which replaces the old Sleep() function.
UI_Sleep() calls the GTK+ UI update function while it's sleeping, so
the menus don't get laggy.

Bug: When Gens is sleeping for large periods of time (e.g. if no ROM
is loaded and the intro style is set to 0), SDL doesn't refresh that
often, which can lead to gray damage areas in the Gens window while
navigating the menus. To fix this, set the GTK+ window background to
black. (Keeping it as gray for now to debug other problems.)