Added the C64 character set.
This font will be used to draw messages and FPS directly to the
SDL, GL, or DDraw framebuffer, instead of drawing to MD_Screen.

TODO: Fill in all the missing characters. The entire 0x80-0xFF
range is missing, and some characters in the 0x00-0x7F range
are missing. Most of the important characters are there, though.