Removed CCnet support.
CCnet was a module to download ROMs from Console Classix. There didn't
seem to be any code in Gens/Linux that actually made use of it, although
there is code in Gens/Win32 that does use it. Regardless, Console Classix
is a rather pointless site that seems to be illegally selling subscriptions
for ROM images. So, I removed all references to CCnet as well as all the
support code.