Cell_8x8_Dump(): Fixed incorrect pattern colors in 32-bit color mode.
Specifically, if a palette other than the first palette (palette 0)
was selected, the colors would show up incorrectly. This was caused
by an incorrect "shl ebp, 1", which was expecting an "shl ebp, 5"
to be executed before it. The "shl ebp, 5" isn't executed because
the jump to .32BIT is done before the shift, so the "shl ebp, 1"
needed to be changed to "shl ebp, 6".