[Win32] Implemented the File, Change State submenu.
Note that Windows doesn't have an easy way to obtain handles of
individual menu items without saving them, so the menus have to
be deleted and recreated. So, the currently selected savestate
is marked as checked in the Create function.

(Actually, this might improve performance on the GTK+ version, so
I may redo the GTK+ version later on.)