[Win32] gens_window: Set Window_Pos if the window is moved.
This saves the current window position, which is used when the program
is loaded so it will show up in the same position.

TODO: This is currently unused in the GTK+ version. See if it's worth
adding, or if it should just remain Win32 only, since Linux WMs are
usually smarter at window positioning than Windows and thus don't
need this sort of hack.