Input_DInput::getKey(): Only use DirectInput keys.
There's really no point in using Windows Virtual Keys here, since input
is all done with DirectInput. Notably, the backslash key is VK_OEM_5
using Virtual Keys, but it's '\\' with DirectInput. And of course, this
is only true on the US English layout; on other layouts, backslash is
a different Virtual Key.