[Win32] Initial implementation of the "Country Code Order" window.
Some changes in other parts of the code had to be made in order to
accomodate drawing icons on buttons in different versions of Windows.
Specifically, on XP, using BM_SETIMAGE removed the visual style.
Instead, BUTTON_IMAGELIST and BCM_SETIMAGELIST (which isn't available
in the WINE headers yet and is only available in XP and later) has to
be used. So, I added a conditional that uses BUTTON_IMAGELIST on XP
and later and BM_SETIMAGE on Windows 2000 and earlier.

Unfortunately, WINE does not support BUTTON_IMAGELIST and BCM_SETIMAGELIST
yet, so the WINE version needs to be set to Windows 2000 for the icons to
show up properly.