[Win32] Fixed a bug with DirectInput cooperative levels.
When the Controller Configuration window sets the cooperative level,
the Gens window loses DirectInput, so controls won't work anymore.
To fix this, I added a call to input->setCooperativeLevel() in the
WM_ACTIVATE handler of both the Gens window and the Controller
Configuration window, which is triggered when the window is activated.