Input_DInput: Fixed DirectInput joystick support.
DirectInput will only acquire a joystick if the window that is attempting
to acquire it is visible. That's rather dumb.

TODO: Set the cooperative level to DISCL_BACKGROUND to allow the Gens
window to accept joystick input even when it's not the active window.
Alternatively, make it a user-configurable option. Also, figure out
how to configure this for SDL. Currently, SDL always accepts joystick
input, even when it's not the active window.