save.cpp: Don't save SRAM or BRAM if the ROM filename is empty.
This usually occurs if SegaCD mode is activated and no CD-ROM is
inserted. There's no point to saving the SRAM or BRAM with no
CD-ROM inserted, anyway. Also, the changes in commit
8b95c1cd386392b09385db73eb059d6edbbc6c22 would cause the filename
to be ".brm" or ".srm" instead of "No Disc.brm" or "No Disc.srm",
which on Linux is interpreted as hidden files due to the leading
period.