imageutil.cpp: Converted the screenshot functions to a static class, ImageUtil.
Among other things, the BMP functionality is now compiled regardless of whether
PNG is enabled or not. This will allow for user-configurable screenshot formats
in the future.

Additionally, ImageUtil can be extended later on for saving an arbitrary bitmap
in memory, which may be used for saving a preview screenshot in savestates.