[GTK+] general_options: Changed the colored radio buttons to look like regular buttons.
Some GTK+ themes, including the default theme used on Ubuntu, do not
allow changing the fill color of radio buttons. (Actually, it seems
like allowing that change is a bug in the GTK-Qt Engine.) Setting
the radio buttons to look like regular buttons fixes this problem on
all themes.

Also, changed the color arrays to two-dimensional instead of having
a single one-dimensional array and doing a bunch of calculations
throughout the code to determine the correct array offset.