Added MacOS X compatibility fixes from Sonicblur.
In particular, this redefines the symbol aliases as weak aliases, so they
won't override other symbols. This is needed on OS X because OS X adds an
underscore prefix to symbol names, whereas Linux (and other ELF platforms)
do not.

Also, disabled compilaton of ui/charset.cpp on platforms other than Win32.
It results in a compile error on OS X, and is only needed on Win32 anyway.

TODO:
- mem_s68k.asm needs a fix for OS X.
  - This could be done by specifying a preprocessor directive to indicate
    an ELF platform versus other platforms.
- Output assembled binaries as Mach-O object files on OS X. (-f macho)