gens_menu.c: New menu definitions file.
The new menu definitions file contains a bunch of structs that contain all of
the menu items. This will be used by a new menu parser instead of manually
calling the appropriate menu functions for every single menu item. The main
benefit is that the menus for both the Linux and Win32 versions can be edited
at the same time.