gens_menu.c: Removed GMF_ICON_*.
These were only needed for the GTK+ UI to indicate if an icon was being used
from the stock GTK+ icon set or if an icon was being used from the Gens pixmaps
directory. Since this logic has been moved to the GTK+ Gens window, it is no
longer needed in gens_menu.c.

All GMF_ICON_* constants were removed. To indicate that no icon is in use,
set the "icon" variable to 0.

All GMF_ITEM_* constants were right-shifted by 4.