pwm.c: Various PWM fixes.
Among other things, removed the "srcPointer" crap. This was a bug
on Gens Rerecording's part, and wasn't in the original asm.

The title music in "BC Racers" had static on the right channel, and
After Burner 32X had static on both channels. Unfortunately, both
games are still unplayable due to other bugs in Gens' 32X emulation.