pasre.cpp: Correctly detect absolute pathnames on Win32.
parseArgs() was only checking for an absolute pathname by checking if
the first character was '/', which is correct for Unix, but wrong for
Win32. It now checks for absolute pathnames on Win32 by checking if
the first character is a letter (drive letter), the second character is
a colon, and the third character is a backslash; or, if the first two
characters are both backslashes (UNC path).