Sync_Gens_Window_FileMenu(): Don't synchronize the ROM History submenu.
Synchronizing the ROM History submenu requires opening all the files in
the ROM History list to check the filetypes, which can take a long time,
especially when some of the files are on network shares.

TODO: Store the filetype in the configuration file so Gens doesn't need
to check the files every time.