blit1x_32.asm: Added a 32-bit color 1x renderer.
This is mostly just a port of the 16-bit color renderer. The only major
changes are the sizes of the surfaces and the number of bytes to copy.

Benchmarks: [debug build]
1,000 iterations of the C++ version [32-bit]: 0.072s
1,000 iterations of the asm version [32-bit]: 0.072s

It doesn't seem that an asm-optimized version really helps here.
Perhaps MMX might improve things.

TODO: Port the MMX version.