mdp_render_hq2x_lut.c: Split the hq2x lookup tables into a separate file.
This will allow hq3x and hq4x to share the same lookup tables.
Also, a reference counter was added. Each plugin increments the
reference counter on initialization. When the plugin is shut down,
the reference counter is decremented; if it hits 0, the plugin will
free the allocated lookup tables.