VDraw_*: m_rInfo.bpp should be equal to bppMD, not bppOut.
If a rendering plugin has MDP_RENDER_FLAG_SRC16DST32 set and the color
depth is set to 32, bppOut is 32, but the surfaces given to the rendering
function are both 16-bit. bppMD is 16 in this case, which is the correct
color depth to pass to the rendering plugin in m_rInfo.