[GTK+] about_window: Improved "ice" drawing.
In particular, the GDK pixbuf is allocated when the About window is created,
instead of allocating it and freeing it once per frame. Also, the Gens logo
is only loaded if "ice" is disabled. Previously, it was always loaded,
regardless of whether or not it was going to be used.