ROM::freeROM(): Clear the border color palette entries.
I noticed a bug where if Sonic 2 is loaded, and a 32X ROM is loaded when
Sonic 2's "SEGA" screen is visible, but the 32X firmware isn't available,
the ROM will be closed, but the border color will still be visible.

This ensures that the border color is always reset to black when a ROM
is closed.