PluginMgr::init(): Check the plugins' required CPU flags.
Plugins that require CPU flags that aren't supported by the CPU
will not be loaded.

TODO: Add these plugins to a list of "incompatible" plugins, so they
will show up in the Plugin Manager UI as requiring a CPU with support
for features that aren't supported on the current system.