mdp_render_scanline_50_*.asm: Removed frame pointers.
Currently, many of the asm renderers have text relocations, since they
reference variables from the .data or .rodata section of the file.
With PIC, they need to use the ELF Global Offset Table; otherwise, the
dynamic loader has to load separate copies of the program code in memory
for each program that uses it, and manually adjust the addresses.

In order to fix this problem, we need to free up the %ebx register.
The %ebp register is only used to access the function parameters, so
we can remove the frame pointer and free up the %ebp register and use
that instead of %ebx. This limits debugging a bit, but isn't really
much of a problem, since the renderers "just work".