ROM::Recent_ROMs: Converted to a deque containing structs.
Among other things, the ROM types are now cached in the configuration
file. This removes the need to open every ROM when Gens loads and when
a ROM is opened in order to determine the ROM type. It should decrease
load time significantly, especially when some ROMs in the ROM History
are located on a network share.

Also, changed the "ROMType" enum to a 32-bit bitfield. The low byte
is the system, and the second byte contains various flags, such as
interleaved or BIN/CUE.