Changed the calling conventions for events.
Instead of having separate function declaration for each type of event,
a single declaration is used. This declaration has parameters indicating
the event ID and a pointer to a struct contianing the actual event
information. The pointer must be casted to the appropriate struct type.
Alternatively, some events may not use this field, and may set it to NULL.