input_sdl.c: Optimized input_sdl_gdk_to_sdl_keyval().
GDK and SDL use the same key values for keys that map to ASCII symbols.
Keys that don't map to ASCII symbols in GDK have 0xFF00 set, so they
can be identified more easily.

Also, make sure uppercase keys are converted to lowercase. (Incidentally,
this may fix a bug I found before: Shift-A wasn't recognized by the SDL
input handling routine, and it's because GDK reported Shift-A as 0x41
[uppercase] instead of 0x61 [lowercase]!)