src/mdp/misc/*/Makefile.am: Added -lgdi32 to LIBADD.
Linking to gdi32.dll is required in order to use GDI functionality,
such as CreateFontIndirect().

NOTE: In order to compile this using a Linux cross-compiler, the
original gdi32.dll is required in /usr/i686-mingw32/usr/lib/libgdi32.dll.
The Wine version will not work, since that's actually an ELF executable.
I tested this with the Windows NT 4.0 gdi32.dll, and it worked.