Added a new easter egg: "Shake Cartridge"
If Ctrl-Backspace is pressed while an MD ROM is loaded, and the program
counter is currently at an area in the ROM, that area of the ROM will
temporarily be inverted. This simulates shaking the cartridge on an
actual MD. Note that this isn't implemented for 32X or SegaCD games.