mdp_host.h: Removed the MDP_PTR_RGB16toYUV pointer.
This lookup table is only used by the hq?x renderers. The lookup table
code has been moved back to those renderers. The downside is that each
of those renderers will use 256 KB of its own RAM for the lookup tables
instead of sharing RAM with the emulator, but the upside is that the
emulator author doesn't need to implement support for this lookup table.

LUT16to32 is still implemented in MDP for now, since it's useful for
supporting plugins that only support 16-bit color.

TODO: The RGB16toYUV code in the hq?x renderers uses mprotect() to mark
the allocated lookup tables as read-only. Make sure this works properly
on Win32 and other environments.