Converted the built-in renderers (Normal, Double) from Intel/NASM syntax to GNU/AT&T syntax.
This reduces the dependency on NASM, which unfortunately has some
compatibility issues between different versions. For instance, an
asm file that assembles fine with NASM 2.05.01 may result in a
phase error with NASM 2.04.

Also, GNU as provides better macros than NASM, since it uses the
C preprocessor. It predefines things like __linux__ and __ELF__.
And I like the "%" prefix for registers. :)