input.c: Added a feature to disable input restrictions.
When disabled, this allows both U+D or L+R to be pressed at the same time.
This is possible on the real system if you disassemble the controller and/or
use a poorly-designed third-party controller.

There is a new gens.cfg setting for this feature, "Restrict Input", located
in the [Input] section. If 1, input is restricted; if 0, input is unrestricted.