io_teamplayer.c: Improved Teamplayer compatibility.
I borrowed some code from Genesis Plus to improve Teamplayer compatibility.
Also, set the initial controller state to 0x60, which is required in order
for NBA Jam TE to work with the Teamplayer adapter properly.

Also, added a function to initialize controllers. Init_Controllers() is
called in the system init and reset functions.