configure.in: Check for strdup().
Added a new macro file, compat_m.h, which contains compatibility macros
for systems that don't support certain functions. The first function in
this file is gens_strdup(), which acts as a wrapper for strdup() on
systems that don't support it.

Also, eliminated a potential memory leak in the Win32 version of
gens_window_sync.cpp::Sync_Gens_Window_PluginsMenu().