sighandler.c: Signal handler.
If Gens/GS crashes, the signal handler will show a message box with
the signal number and a note asking the user to send a bug report.

I was going to add a glibc backtrace too, but the glibc backtrace
functions don't seem to properly resolve symbols. That, and they're
only available on Linux.