mdZ80: Use the "fastcall" calling convention on Linux instead of regparm(2).
gcc's regparm(2) calling convention passes the first two parameters in %eax and %edx,
whereas Win32's fastcall calling convention uses %ecx and %edx. Switching everything
to fastcall reduces the amount of %ifdef's.

Note that the fastcall attribute was added in gcc-3.4, so Gens will not work
correctly unless it is built with gcc-3.4 or later.