macros/hashtable.hpp: Fixed GENS_HASHTABLE<mdp_t*, [type]> with the GNU hash_map.
hash_map is horribly broken and only works for basic types. The libstdc++
maintainers aren't going to fix this, since TR1 unordered_map is better,
but even if they did fix it, it wouldn't help any. The only reason hash_map
is supported is for gcc-4.0 compatibility on MacOS X (since TR1 unordered_map
is broken on gcc-4.0), and for whatever reason, Apple is still using gcc-4.0,
even on the latest versions of MacOS X.

This bug was found by Puto.