ens_ui_win32.cpp: ICC_STANDARD_CLASSES is WinXP and later only.
MSDN doesn't say this, but the MS and Cygwin headers both restrict
ICC_STANDARD_CLASSES to _WIN32_WINNT >= 0x0501. (Wine's headers don't
restrict it, which should probably be reported as a bug to the Wine
developers.)

MSDN page: http://msdn.microsoft.com/en-us/library/bb775507(VS.85).aspx
(Of course, if you're reading this sometime in the near future, chances
are that the MSDN page link won't be working anymore. The page is about
the INITCOMMONCONTROLSEX structure.)