Use libtool to build all libraries, including static libraries.
This will be useful later on when Gens/GS is split into a frontend
executable and a backend library (libgens.so).

Also, added a dummy C file to Starscream (src/starscream/dummy.c)
to convince libtool to link it properly. The C file doesn't contain
any actual code; it just has a comment indicating that it's a dummy
file for libtool.