Patch 2 (Haiku OS port)
configure.in: Check GTK and SDL macro availability before expansion

This patch checks the availability of the AM_PATH_GTK_2_0 and the
AM_PATH_SDL macros during the running of aclocal before expanding them
since the m4 files for either of those macros may not be present when
compiling on some platforms.

On platforms that do have the required m4 files in /usr/share/aclocal
but not the necessary libs, the usual ./configure errors for those
dependencies will appear.