Haiku Patch 4
Hi again!

This patch helps the Haiku build process along further and introduces
new files for the user interface, they are bare bones at first just to
satisfy the structure for the compiler. The UI is partially written,
sitting in another directory for when everything compiles nicely. I'm
fairly certain the Haiku UI can be easily laid out exactly as they are
for the Windows and GTK interfaces file-wise so I copied over the
headers exactly as they are for now but removed the GTK specifics.

There are #ifdefs introduced in this patch which will be removed very
soon in coming patches. They are just there for now in one file so that
I can get it compiling on Haiku - the minute I have a replacement
implementation for the function #ifdef'd out, the #ifdefs will be gone :-)

After applying the patch, the zip file is a set of new files to be
extracted in the top level gens directory (they extract into src and
overwrite one haiku-specific ui file) and then added with "git add
src/gens/ui/haiku/*"

Thanks!

Phil