Haiku Patch 5 (Initial input code)
Hi David,

This patch contains the next few compile fixes for Haiku - it still
won't compile fully on that OS yet since input is a bit trickier and
needs more thought. The attached zip file is also ready to be unzipped
directly into the "input" folder and git added as part of the commit.

If anything looks back to front, go ahead and switch it around. Adding
those stubs in the last patch was very helpful since I only thought
about doing that after sending the third email heh

Do you have any tips for how input should be handled? Normally in Haiku,
there is a BMessage object which is sent to the BWindow or BApplication
and the keycode is read at that point... it looks like I need to send
those keycodes off to input.h or input.c and handle them there to keep
everthing working the same way but if you are planning to change it
around at all any time soon, let me know and I can make it skip input
while I focus on getting the rest to compile.

Thanks!

Phil.