vdraw.c: Initial backend fallback support.
If SDL+OpenGL is selected but OpenGL couldn't be initialized for some
reason (missing GLX extension, etc), then Gens/GS will revert to the
SDL backend.

This commit introduces a new array, uint8_t vdraw_backends_broken[].
Each element corresponds to a backend ID. If the element is 0, then
that backend is either working or hasn't been tested. If the element
is nonzero, then the backend is broken.