Version bumped all feature-complete external plugins to v1.0.0.
"Feature-complete" means that the plugin is essentially finished for
the initial release, though may change if the MDP interface changes
before the MDP v1.0 release.

Some plugins that aren't yet feature-complete are v0.9.0. This
includes the IPS plugin (needs some polishing) and the NTSC renderer
(needs a decoder matrix option in the GUI).