Haiku Patch 6
Hi David,

This patch adds a little more to the input handling code for Haiku. It
is still organizational for now but outlines the structure I'm planning
to use for input.

Can I request that after applying the patch, the files
input/input_haiku.c and input_haiku.h are renamed in git to
input_haiku.cpp and input_haiku.hpp respectively?

Thanks,

Phil