Haiku Patch 7
Hi again David,

This patch helps the compiler get further still while compiling under
Haiku providing --enable-debugger=no.

With regard to removing the __cplusplus check - I did that in advance of
the file being renamed with the .hpp since it is meant to be compiled as
c++ there specifically so it doesn't matter too much whether those lines
are there or not (as far as I know).

The next step is the vdraw section, then I plan to work on bringing the
GUI elements into line with the GTK version and adding them to the project.

Thanks,
Phil