cpuflags.c: If the OS doesn't support SSE, clear the SSE flags.
Executing an SSE instruction on an operating system that doesn't support
SSE results in an Illegal Instruction error, even if the CPU supports it.

On POSIX-compliant systems, a temporary SIGILL handler is set up and an
SSE instruction is run. If a SIGILL occurs, SSE is disabled; otherwise,
the CPU flags are left as-is.

On Win32, mingw doesn't properly support returning from a SIGILL handler,
so the OS version number is checked. The first versions of Windows that
supported SSE are Windows 98 and Windows 2000, so any version of Windows
with a version number older than 4.10 will have SSE disabled. (I've read
conflicting reports that Windows 98 First Edition supports SSE with a
DirectX upgrade while others say that only Windows 98 Second Edition
supports SSE, so for now I'll just enable SSE for any Windows 98.)