rom.cpp: Added a variable, ROM_ByteSwap_State, to indicate the current byteswap state of loaded ROMs.
This will be used for MDP_EMUCTRL_RELOAD_INFO, since the emulator
will need to reload the information from the ROM header. At ROM
startup, the loaded ROM is not byteswapped; however, after startup,
the ROM is byteswapped (on x86). Thus, it must be byteswapped again
in order to be reloaded properly.