sighandler.c: Improved the output of the signal handler on platforms that support sigaction().
On platforms that support sigaction(), the signal information will be read,
and the signal code will be displayed in addition to the signal. This allows
for more fine-grained error display, such as FPE_INTDIV in cases of integer
division by zero. (SIGFPE by itself doesn't say what the specific error was.)