blargg_ntsc/md_ntsc.cpp: Reimplemented interpolation support.
The interpolation code is in a separate function, which makes it somewhat
easier to maintain.

This commit also reimplements interpolation with scanlines. This should
be the last commit needed to fully reimplement the double-scan functionality
without needing to reread VRAM, which should make it a lot faster when
used on the Win32 DirectDraw 4 renderer.