input_sdl.c: Reworked input_sdl_get_key() to ignore initial states.
This makes it a lot easier to configure some gamepads that feature triggers,
such as the Nintendo GameCube Controller and Nintendo Wii Classic Controller.
Specifically, the trigger axes are set to <-10,000 when not pressed, which
would cause Gens/GS to assume that those axes were "moved".

Now, the current state of the controllers is saved before checking for
changes, and if the current state was "pressed", the new state will be
ignored. This makes it a lot easier to configure these types of controllers.

Note: POV hat support is apparently broken right now. It'll be fixed in
the next commit.