plugin_check.c: Initial plugin verification functions.
Also, reworked some TEST_* macros in main.h to provide varargs and
non-varargs versions, and added TEST_INFO(), which prints information
in purple. This is used for printing the plugin description struct.

Note that right now, mdp->func->init(NULL) is being called, and most
of the Gens/GS plugins do not like having a NULL host_srv pointer
passed to them. This results in a segmentation fault. The plugins
should be updated to return an error code if host_srv is NULL.