audio_dsound.cpp: Fixed a massive lag regression from the "back-to-c" branch.
Commit 7782d084dbd8bac823b79b11d2ac656bbcc5b312 ported the DirectSound code
to a C-like architecture instead of classes. In addition to the general code
restructuring, some of the code was changed. For instance, the code in
Audio_DSound::wpSegWait() [now audio_dsound_wp_seg_wait()] was changed:

-void Audio_DSound::wpSegWait(void)
+void audio_dsound_wp_seg_wait(void)
 {
-	while (WP == getCurrentSeg())
+	while (WP == audio_dsound_get_current_seg())
	{
-		GensUI::sleep(1);
+		Sleep(1);
	}
 }

On Win32, GensUI::sleep() divides the specified number of milliseconds by 5
before actually sleeping. With integer arithmetic, 1 / 5 == 0, so this was
effectively a nop. Sleep(), on the other hand, is a Win32 API function, and
always sleeps, regardless of the specified value. On most systems, this
didn't really cause problems. However, one user in particular, djohe, had an
issue where this caused massive lag, resulting in framerate drops. Simply
removing the Sleep() call fixed the problem, and the framerate is now a
constant 60.0 fps. (50.0 fps for PAL.)

Thanks to djohe from #retro for helping to diagnose and solve this problem.