[Win32] Improved window reinitialization when switching to/from Full Screen.
Wine doesn't like it when the WS_CLIPSIBLINGS style is removed from a
window, and refuses to redraw any part of the window afterwards.
Reapplying it later doesn't fix anything, either. This makes sure that
WS_CLIPSIBLINGS is never removed from the window.

Wine's still having a problem where it won't re-enable window borders
after returning from fullscreen, but it's better than before.

This also merges the reinitialization code in the DirectDraw 4 renderer and
the GDI renderer.