gsft_byteswap.c: Only byteswap the requested amount of bytes.
Commit 6b24d512e78c79a70dca3b35cd403e68082e8ff7 optimized the byteswap code,
but inadvertently made it so it would byteswap twice the amount specified.
For Sonic 2, this wasn't a problem, since it's only 1 MB. Sonic 3D, however,
is 4 MB, so the code ended up byteswapping 8 MB, which caused other memory
to be overwritten. The result was that Starscream crashed.

This commit fixes it so the correct amount of bytes are swapped. Also, the
byteswap function now enforces even lengths, and removes the last byte if
the length is odd.