v_effects.cpp: Fixed the Crazy Effect on Linux.
On MinGW and MSVC, RAND_MAX is defined as 32767. On Linux, RAND_MAX is 2147483647.
The randomizer was checking for (rand() > 0x2C00), and with a massive RAND_MAX,
this would almost always equate to true.

AND'ing rand() with 0x7FFF makes it work like on Windows, so the crazy effect
works properly now.