v_effects.cpp::T_Update_Crazy_Effect(): Only update the visible area of the screen.
This fixes the Crazy Effect when used with some renderers that write to the
source buffer, e.g. 2xSaI / Super Eagle / Super 2xSaI. Ideally, the renderers
should be fixed, but this is a good enough workaround for now.

That, and the workaround probably reduces CPU usage, since it doesn't have
to render to the entire 336x240 buffer.