Final changes to allow full basic compile on Haiku pre-alpha gcc4
Hi David,

This patch (and the attached zip of files) complete the set of changes
necessary in order to compile Gens GS on Haiku GCC4.

The foundations are now in place for the for the main part of the port -
filling in the stubs with Haiku-specific code.

If there is anything which you think is not particularly clean as it is
in gens, or that you plan to rearrange, please let me know, and I can
work on starting out with that in mind and we can avoid extra untangling
work in the future :-)

Regards,
Phil

Signed-off-by: David Korth <gerbilsoft@verizon.net>