[Win32] debug_window.c: Initial debug window support for Gens/GS Win32.
This allows for console output via an optional command-line parameter
in the release version of Gens/GS Win32, which usually has the subsystem
set to GUI instead of Console.

This commit adds the debug window and the infrastructure for enabling it.
There's no command line option to enable it yet; it's enabled all the time.
Also, nothing's actually logged yet.