g_main_win32.cpp: Call DestroyWindow(gens_window) after End_All().
This is an accidental regression from commit 1558532fb72427ced8572af61c4e6e4176572b80.
It broke quitting while in fullscreen, since the Gens window would
be receiving a bunch of messages that weren't handled, causing the
"Empty the message queue" loop to run infinitely. (Sometimes, it
resulted in a crash. Not sure why.).