vdraw_text.cpp: Initial prerendering optimization for text drawing.
The prerendering code looks up the character data in C64_charset[] when
the text is printed to the screen, and puts it in an array that's more
cache-friendly. This should potentially improve performance. (I haven't
actually benchmarked it, though. I probably should!)

Also, fixed a bunch of other performance issues in text drawing, including
aborting if the message length is 0.