vdraw_text.cpp::T_drawText(): Increase text shadow offset for 2x text rendering.
I decreased the offset because it made the text look weird in 1x, but it turns
out that the original offset is needed for 2x. The program now sets the original
offset for 2x and the decreased offset for 1x.

This is a regression introduced in one of the text drawing optimization
commits. I can't figure out which one it was, but I don't really care.
Milestone 6 didn't have this issue.