Better integrate main application loop with UI
Relocate the call to the the main GensApplication loop and correct its
interaction with GensMainLoop().

Hi David, the patch and text for the commit are above. Just thinking out
loud: you may want to move the GensMainLoop code (just the part within
the while loop that I've duplicated into the Haiku file) to another
function in g_main.cpp, that way, we can call just one function from
both GensApplication::Pulse() and within the while loop of
GensMainLoop() when the app needs to "think".

Regards,
Phil