Revert "vdraw_cpp.cpp::vdraw_reset_renderer(): Update the screensaver suspend state."
This reverts commit b53c6b187b6851a3deb0bde682b1bfeb8301ed58.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.