Revert "Call GensUI::updateScreenSaverSuspend() after Escape or Pause is pressed."
This reverts commit b277360778e48cf0b444de9af30a62596413e7f2.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.