Revert "GensUI::updateScreenSaverSuspend(): Initial logic to determine if a game is running or not."
This reverts commit 3627f9485ae937d1f90eb3f0066c48cb09f2c6c2.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.