Revert "genopt_window.cpp: Call GensUI::updateScreenSaver() after saving settings."
This reverts commit 31eb84d61cdea5787fc01ccb5b9db23f196129a9.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.