Revert "[Win32] genopt_window.cpp: Added a checkbox for "Suspend Screensaver"."
This reverts commit cf1afda8e15553100fa62bbf6b0356ea4003c706.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.