Revert "[GTK+] genopt_window.cpp: Added a checkbox for "Suspend Screensaver"."
This reverts commit fdac02dbf8756263a7fa35b0a348badd09d3a0f2.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.