Revert "GensUI: Initial stubs for disabling the screensaver while a game is running."
This reverts commit 0a863ba0ff8f6d3ffecb7e2a46899bd90b7dc6f3.

None of the X11 screensavers I've tried (XScreenSaver, KDE4, GNOME) listen to
any of the X11 screensaver protocols (either the standard X11 screensaver
protocols or the XScrnSaver protocol.) I have no idea why, either.

SDL disables the screensaver on startup by default anyway, so it's not that
big of a deal. It would be nice if I could re-enable the screensaver when a
game wasn't running, though.

Conflicts:

	src/gens/emulator/g_main.hpp