input_sdl.c: Fixed POV hat detection in input_sdl_get_key().
For whatever reason, it was reading from an uninitialized SDL_Event struct.
It seemed to work when I first implemented it, but it stopped working
sometime after.

Anyways, the solution is to get rid of the SDL_Event struct and use the
correct variables in the INPUT_GETKEY_POVHAT_DIRECTION() macro.