md_palette.cpp: Added a color scaling option.
Color scaling is needed in order to brighten the colors to make them
more accurate to the original console. Simply left-shifting the original
MD color values by 4 (for 32-bit color) doesn't work, since the MD
programming manual specifies that a CRAM color component value of 0xE
is full brightness.

The color scaling algorithm first multiples each component by the maximum
value of a color component in the current color depth (0x1F, 0x3F, 0xFF),
and then it divides the result by the maximum value of a color component
in the current color depth if it were to use simple bit-shifting to scale
the colors (0x1C, 0x38, 0xE0). This accurately scales 0xE to 224 in 32-bit
color, for example. It also works for other values.

TODO:
- 32X palette scaling.
- Win32 UI for palette scaling. (Graphics, Color Adjust)
  - GTK+ UI is already implemented.
- (later?) MDP render flag to indicate if a plugin does scaling
  by itself. Blargg's NTSC filter does this.
  - Might not be needed, since I didn't notice any difference with
    scaling on or off when using the NTSC filter. This is probably because
    AND'ing the scaled value (32-bit) with 0xE0E0E0 results in the unscaled
    value, which is probably what the NTSC filter does.

This option was suggested by Hivebrain on Sonic Retro.